Now that we have our Module type, let's go back to our Track type.
We know that a Track can have multiple modules, so let's add a field to Track called modules, which returns an array of Module objects. This array can't be null so we add an exclamation point at the end, and the entries in the array can't be null either, so we add another exclamation point after Module.