We've used the third argument context before, to access the dataSources key. This is where our trackAPI.getTrack method lives. This method is expecting the id of a track. We'll be specifying that value later as an argument when we build the query.
To access this id, we can use the second parameter in the resolver: args. args is an object that contains all GraphQL arguments that were provided for the field. We can destructure this object to access the id property, then pass that id into our getTrack method.

Then, we can return the results of that method.