Now let's build the skeleton of the Track page.
We'll declare a functional component that takes a trackId destructured from the props input. This prop will be coming as a parameter from the route, or the browser's URL, which we'll set up later.

For now, let's return the layout of the page. Make sure to export the component below.