We've set up our data source to make a PATCH call to our REST API. Now it's time to write resolvers to put our data source to work.
In the server/src folder, let's open up the resolvers.js file. We need to add another entry to our resolvers object for our mutation.

We can add it just below our Query resolvers, and we need to give it the same name as in our schema, which is Mutation (note the capital M!).