So far, we've always immediately returned the results of our TrackAPI calls, because the results match the structure required by our schema.
However, in this case, we have 3 fields required by our schema that are not present in the returned response: code, success and message. This makes sense because they are tied to the REST operation status itself, so we need to set those 3 based on that status.

First, let's wait for the TrackAPI call to finish returning the data. We can do this by making the function async, then storing the results in a variable called track, making sure to await the results.