Then, we return an object with the same properties as the object in the success case, but with different values.
We could set the code to be 404, but we can also be more dynamic and use the values that Apollo Server and the RESTDataSource class provide. When an error occurs, Apollo Server attaches an extensions field to that error that contains relevant error details.

In this case, as our TrackAPI extends RESTDataSource, this extensions object will be enriched with a response property, which provides some additional information about the HTTP response itself. We can return the status property, which refers to the HTTP status code. Perfect to use here!