Now, here's a twist: unlike with useQuery, calling useMutation doesn't actually execute the mutation automatically!
Instead, the useMutation hook returns an array with two elements, which we'll start to destructure here.

The first element is the mutate function we'll use to actually run the mutation later on. We'll call it incrementTrackViews. The second element is an object with information about the mutation: loading, error and data. This component doesn't need it, so we don't have to extract it.